LazyBailOutRecord
