+++ chakracore-cxx/lib/Backend/Encoder.h
@@ -4,6 +4,7 @@
 //-------------------------------------------------------------------------------------------------------
 #include "EncoderMD.h"
 #include "Backend.h"
+#include "LazyBailOutRecord.h"
 
 ///---------------------------------------------------------------------------
 ///
